Only changes to the names of the readers to fit the convention set in the rest of the package with detector at start of name.
