+++ FoGSE/playback_readers/CdTeHKPlaybackReader.py
@@ -9,7 +9,7 @@ Can read:
 import struct
 import numpy as np
 
-from FoGSE.read_raw_to_refined.readRawToRefinedBase import ReaderBase
+from FoGSE.readers.BaseReader import BaseReader
 
 from FoGSE.readBackwards import BackwardsReader
 from FoGSE.parsers.CdTeparser import CdTecanisterhkparser
@@ -17,7 +17,7 @@ from FoGSE.parsers.CdTeparser import CdTecanisterhkparser
 from FoGSE.collections.CdTeHKCollection import CdTeHKCollection
 
 
-class CdTeHKReader(ReaderBase):
+class CdTeHKReader(BaseReader):
     """
     Reader for the FOXSI CdTe instrument.
     """
@@ -32,7 +32,7 @@ class CdTeHKReader(ReaderBase):
         if datafile is None:
             return
 
-        ReaderBase.__init__(self, datafile, parent)
+        BaseReader.__init__(self, datafile, parent)
 
         self.define_buffer_size(size=796)#100_000#32_780
         self.call_interval(100)
