+++ FoGSE/readers/CMOSHKReader.py
@@ -6,14 +6,14 @@ Can read:
     * CMOS PC
 """
 
-from FoGSE.read_raw_to_refined.readRawToRefinedBase import ReaderBase
+from FoGSE.readers.BaseReader import BaseReader
 
 from FoGSE.readBackwards import BackwardsReader
 import FoGSE.parsers.CMOSparser as cmosp
 from FoGSE.collections.CMOSHKCollection import CMOSHKCollection
 from FoGSE.utils import get_frame_size, get_system_value
 
-class CMOSHKReader(ReaderBase):
+class CMOSHKReader(BaseReader):
     """
     Reader for the FOXSI CMOS instrument.
     """
@@ -28,10 +28,10 @@ class CMOSHKReader(ReaderBase):
         if datafile is None:
             return
         
-        ReaderBase.__init__(self, datafile, parent)
+        BaseReader.__init__(self, datafile, parent)
         
         self.define_buffer_size(size=get_frame_size("cmos1", "hk")) # 536 bytes
-        self.call_interval(get_system_value("gse", "display_settings", "cmos", "hk", "read_raw_to_refined", "read_interval"))
+        self.call_interval(get_system_value("gse", "display_settings", "cmos", "hk", "readers", "read_interval"))
 
     def extract_raw_data(self):
         """
