+++ FoGSE/readers/CMOSQLReader.py
@@ -3,19 +3,19 @@ Create a class that will read the LOG file containing raw binary data received f
 FOXSI and parse the data to be readyfor the GUI plotting windows. 
 
 Can read:
-    * Timepix
+    * CMOS
 """
 
-from FoGSE.read_raw_to_refined.readRawToRefinedBase import ReaderBase
+from FoGSE.readers.BaseReader import BaseReader
 
 from FoGSE.readBackwards import BackwardsReader
-from FoGSE.parsers.Timepixparser import timepix_parser
-from FoGSE.collections.TimepixCollection import TimepixCollection
+from FoGSE.parsers.CMOSparser import QLimageData 
+from FoGSE.collections.CMOSQLCollection import CMOSQLCollection
 from FoGSE.utils import get_frame_size, get_system_value
 
-class TimepixReader(ReaderBase):
+class CMOSQLReader(BaseReader):
     """
-    Reader for the FOXSI Timepix instrument.
+    Reader for the FOXSI CMOS instrument.
     """
 
     def __init__(self, datafile, parent=None):
@@ -24,10 +24,10 @@ class TimepixReader(ReaderBase):
         Parsed : human readable
         Collected : organised by intrumentation
         """
-        ReaderBase.__init__(self, datafile, parent)
-        
-        self.define_buffer_size(size=5) # bytes, get_frame_size("timepix", "tpx")
-        self.call_interval(get_system_value("gse", "display_settings", "timepix", "tpx", "read_raw_to_refined", "read_interval"))
+        BaseReader.__init__(self, datafile, parent)
+        # The magic number for CMOS PC data is 590,848. The magic number for CMOS QL data is 492,544.
+        self.define_buffer_size(size=get_frame_size("cmos1", "ql"))
+        self.call_interval(get_system_value("gse", "display_settings", "cmos", "ql", "readers", "read_interval"))
 
     def extract_raw_data(self):
         """
@@ -39,11 +39,11 @@ class TimepixReader(ReaderBase):
         `list` :
             Data read from `self.data_file`.
         """
-        return self.extract_raw_data_timepix()
+        return self.extract_raw_data_cmos()
     
-    def extract_raw_data_timepix(self):
+    def extract_raw_data_cmos(self):
         """
-        Method to extract the CdTe data from `self.data_file` and return the 
+        Method to extract the CMOS data from `self.data_file` and return the 
         desired data.
 
         Returns
@@ -55,15 +55,14 @@ class TimepixReader(ReaderBase):
         # forward=True: reads buffer from the back but doesn't reverse the data 
         try:
             with BackwardsReader(file=self.data_file, blksize=self.buffer_size, forward=True) as f:
-                datalist = f.read_block()
-                
-            if self._old_data==datalist:
+                data = f.read_block()
+            if self._old_data==data:
                 return self.return_empty() 
         except FileNotFoundError:
             return self.return_empty() 
         
-        self._old_data = datalist
-        return datalist
+        self._old_data = data
+        return data
 
     def raw_2_parsed(self, raw_data):
         """
@@ -78,17 +77,17 @@ class TimepixReader(ReaderBase):
         Returns
         -------
         `tuple` :
-            Output from the CdTe parser.
+            Output from the CMOS parser.
         """
         # return or set human readable data
         # do stuff with the raw data and return nice, human readable data
         try:
-            tot, flx, flgs = timepix_parser(raw_data)
+            linetime, gain, exposure_pc, pc_image = QLimageData(raw_data)
         except ValueError:
             # no data from parser so pass nothing on with a time of -1
             print("No data from parser.")
-            tot, flx, flgs = (None,None,None)
-        return tot, flx, flgs
+            linetime, gain, exposure_pc, pc_image = (-1,None,None,None)
+        return linetime, gain, exposure_pc, pc_image
 
     def parsed_2_collection(self, parsed_data):
         """
@@ -97,17 +96,16 @@ class TimepixReader(ReaderBase):
         Parameters
         ----------
         parsed_data : `tuple`
-            Output from the CdTe parser.
+            Output from the CMOS parser.
 
         Returns
         -------
-        `FoGSE.detector_collections.CdTeCollection.CdTeCollection` :
-            The CdTe collection.
+        `FoGSE.detector_collections.CMOSQLCollection.CMOSQLCollection` :
+            The CMOS collection.
         """
         # take human readable and convert and set to 
         # CdTeCollection(), TimePixCollection(), CMOSCollection()
-        col = TimepixCollection(parsed_data, 0)#self.old_data_time) #replace the old data time with 0 to allow even old data trhough if it gets to this stage (come back to this!)
-
-        # if col.last_data_time>self.old_data_time:
-        #     self.old_data_time = col.last_data_time
+        col = CMOSQLCollection(parsed_data, self.old_data_time)
+        if col.last_data_time>self.old_data_time:
+            self.old_data_time = col.last_data_time
         return col
