+++ FoGSE/readers/CdTePCReader.py
@@ -8,16 +8,17 @@ Can read:
 
 import struct
 import numpy as np
+import os
 
-from FoGSE.read_raw_to_refined.readRawToRefinedBase import ReaderBase
+from FoGSE.readers.BaseReader import BaseReader
 
 from FoGSE.readBackwards import BackwardsReader
-from FoGSE.parsers.CdTeparser import CdTedehkparser
-# from FoGSE.parsers.CdTeframeparser import CdTerawdataframe2parser
-from FoGSE.collections.DECollection import DECollection
+from FoGSE.parsers.CdTeparser import CdTerawalldata2parser
+from FoGSE.parsers.CdTeframeparser import CdTerawdataframe2parser
+from FoGSE.collections.CdTeCollection import CdTeCollection
 from FoGSE.utils import get_frame_size, get_system_value
 
-class DEReader(ReaderBase):
+class CdTePCReader(BaseReader):
     """
     Reader for the FOXSI CdTe instrument.
     """
@@ -28,14 +29,10 @@ class DEReader(ReaderBase):
         Parsed : human readable
         Collected : organised by intrumentation
         """
-
-        if datafile is None:
-            return
-
-        ReaderBase.__init__(self, datafile, parent)
+        BaseReader.__init__(self, datafile, parent)
         
-        self.define_buffer_size(size=get_frame_size("cdtede", "hk")) # 32 bytes
-        self.call_interval(get_system_value("gse", "display_settings", "cdtede", "hk", "read_raw_to_refined", "read_interval"))
+        self.define_buffer_size(size=get_frame_size("cdte1", "pc")) # 32_780 bytes
+        self.call_interval(get_system_value("gse", "display_settings", "cdte", "pc", "readers", "read_interval"))
 
     def extract_raw_data(self):
         """
@@ -47,9 +44,9 @@ class DEReader(ReaderBase):
         `list` :
             Data read from `self.data_file`.
         """
-        return self.extract_raw_data_cdtehk()
+        return self.extract_raw_data_cdte()
     
-    def extract_raw_data_cdtehk(self):
+    def extract_raw_data_cdte(self):
         """
         Method to extract the CdTe data from `self.data_file` and return the 
         desired data.
@@ -63,14 +60,18 @@ class DEReader(ReaderBase):
         # forward=True: reads buffer from the back but doesn't reverse the data 
         try:
             with BackwardsReader(file=self.data_file, blksize=self.buffer_size, forward=True) as f:
-                data = f.read_block()
-            if self._old_data==data:
+                iterative_unpack=struct.iter_unpack("<I",f.read_block())
+                datalist=[]
+                for _,data in enumerate(iterative_unpack):
+
+                    datalist.append(data[0])
+            if self._old_data==datalist:
                 return self.return_empty() 
         except FileNotFoundError:
             return self.return_empty() 
         
-        self._old_data = data
-        return data
+        self._old_data = datalist
+        return datalist
 
     def raw_2_parsed(self, raw_data):
         """
@@ -90,12 +91,12 @@ class DEReader(ReaderBase):
         # return or set human readable data
         # do stuff with the raw data and return nice, human readable data
         try:
-            parsed_data, error_flag = CdTedehkparser(raw_data) #CdTerawalldata2parser(raw_data)# 
+            flags, event_df, all_hkdicts = CdTerawdataframe2parser(raw_data) #CdTerawalldata2parser(raw_data)# 
         except ValueError:
             # no data from parser so pass nothing on with a time of -1
             print("No data from parser.")
-            parsed_data, error_flag = ({"status": "N/A","ping": "N/A","temp": "N/A","cpu": "N/A","df_GB": "N/A","unixtime": "N/A"},None)
-        return parsed_data, error_flag
+            flags, event_df, all_hkdicts = (None,{'ti':np.array([-1]), 'unixtime':np.array([-1]), 'hitnum_al':np.array([-1]), 'hitnum_pt':np.array([-1])},None)
+        return flags, event_df, all_hkdicts
 
     def parsed_2_collection(self, parsed_data):
         """
@@ -113,9 +114,10 @@ class DEReader(ReaderBase):
         """
         # take human readable and convert and set to 
         # CdTeCollection(), TimePixCollection(), CMOSCollection()
-        col = DECollection(parsed_data, 0)#self.old_data_time) #replace the old datat time with 0 to allow even old data trhough if it gets to this stage (come back to this!)
+        col = CdTeCollection(parsed_data, 0)#self.old_data_time) #replace the old datat time with 0 to allow even old data trhough if it gets to this stage (come back to this!)
         # print("Old data time: ",self.old_data_time)
         # print("Newest data time:",col.last_data_time)
         if col.latest_data_time>self.old_data_time:
             self.old_data_time = col.latest_data_time
         return col
+
