+++ FoGSE/readers/DEReader.py
@@ -3,21 +3,18 @@ Create a class that will read the LOG file containing raw binary data received f
 FOXSI and parse the data to be readyfor the GUI plotting windows. 
 
 Can read:
-    * CdTe
+    * DE
 """
 
-import struct
-import numpy as np
-
-from FoGSE.read_raw_to_refined.readRawToRefinedBase import ReaderBase
+from FoGSE.readers.BaseReader import BaseReader
 
 from FoGSE.readBackwards import BackwardsReader
-from FoGSE.parsers.CdTeparser import CdTecanisterhkparser
+from FoGSE.parsers.CdTeparser import CdTedehkparser
 # from FoGSE.parsers.CdTeframeparser import CdTerawdataframe2parser
-from FoGSE.collections.CdTeHKCollection import CdTeHKCollection
+from FoGSE.collections.DECollection import DECollection
 from FoGSE.utils import get_frame_size, get_system_value
 
-class CdTeHKReader(ReaderBase):
+class DEReader(BaseReader):
     """
     Reader for the FOXSI CdTe instrument.
     """
@@ -32,10 +29,10 @@ class CdTeHKReader(ReaderBase):
         if datafile is None:
             return
 
-        ReaderBase.__init__(self, datafile, parent)
-
-        self.define_buffer_size(size=get_frame_size("cdte1", "hk")) # 796 bytes
-        self.call_interval(get_system_value("gse", "display_settings", "cdte", "hk", "read_raw_to_refined", "read_interval"))
+        BaseReader.__init__(self, datafile, parent)
+        
+        self.define_buffer_size(size=get_frame_size("cdtede", "hk")) # 32 bytes
+        self.call_interval(get_system_value("gse", "display_settings", "cdtede", "hk", "readers", "read_interval"))
 
     def extract_raw_data(self):
         """
@@ -90,11 +87,11 @@ class CdTeHKReader(ReaderBase):
         # return or set human readable data
         # do stuff with the raw data and return nice, human readable data
         try:
-            parsed_data, error_flag = CdTecanisterhkparser(raw_data) #CdTerawalldata2parser(raw_data)# 
+            parsed_data, error_flag = CdTedehkparser(raw_data) #CdTerawalldata2parser(raw_data)# 
         except ValueError:
             # no data from parser so pass nothing on with a time of -1
             print("No data from parser.")
-            parsed_data, error_flag = ({"status": "N/A","hv_exec": "N/A","hv_set": "N/A","frame_count": "N/A"},None)
+            parsed_data, error_flag = ({"status": "N/A","ping": "N/A","temp": "N/A","cpu": "N/A","df_GB": "N/A","unixtime": "N/A"},None)
         return parsed_data, error_flag
 
     def parsed_2_collection(self, parsed_data):
@@ -113,7 +110,7 @@ class CdTeHKReader(ReaderBase):
         """
         # take human readable and convert and set to 
         # CdTeCollection(), TimePixCollection(), CMOSCollection()
-        col = CdTeHKCollection(parsed_data, 0)#self.old_data_time) #replace the old datat time with 0 to allow even old data trhough if it gets to this stage (come back to this!)
+        col = DECollection(parsed_data, 0)#self.old_data_time) #replace the old datat time with 0 to allow even old data trhough if it gets to this stage (come back to this!)
         # print("Old data time: ",self.old_data_time)
         # print("Newest data time:",col.last_data_time)
         if col.latest_data_time>self.old_data_time:
