+++ FoGSE/readers/RTDReader.py
@@ -3,18 +3,16 @@ Create a class that will read the LOG file containing raw binary data received f
 the RTDs
 """
 
-# `from PyQt6 import QtCore`
-
-from FoGSE.read_raw_to_refined.readRawToRefinedBase import ReaderBase
+from FoGSE.readers.BaseReader import BaseReader
 
 from FoGSE.readBackwards import BackwardsReader
-from FoGSE.parsers.Powerparser import adcparser
-from FoGSE.collections.PowerCollection import PowerCollection
+from FoGSE.parsers.RTDparser import rtdparser
+from FoGSE.collections.RTDCollection import RTDCollection
 from FoGSE.utils import get_frame_size, get_system_value
 
-class PowerReader(ReaderBase):
+class RTDReader(BaseReader):
     """
-    Reader for the Power readout.
+    Reader for the RTD readout.
     """
 
     def __init__(self, datafile, parent=None):
@@ -23,10 +21,10 @@ class PowerReader(ReaderBase):
         Parsed : human readable
         Collected : organised by intrumentation
         """
-        ReaderBase.__init__(self, datafile, parent)
+        BaseReader.__init__(self, datafile, parent)
         
-        self.define_buffer_size(size=get_frame_size("housekeeping", "pow")) # 38 bytes
-        self.call_interval(get_system_value("gse", "display_settings", "housekeeping", "pow", "read_raw_to_refined", "read_interval"))
+        self.define_buffer_size(size=get_frame_size("housekeeping", "rtd")*2) # 84 bytes
+        self.call_interval(get_system_value("gse", "display_settings", "housekeeping", "rtd", "readers", "read_interval"))
 
     def extract_raw_data(self):
         """
@@ -81,16 +79,8 @@ class PowerReader(ReaderBase):
         """
         # return or set human readable data
         # do stuff with the raw data and return nice, human readable data
-        try:
-            output, error_flag = adcparser(raw_data)
-        except ValueError:
-            # no data from parser so pass nothing on with a time of -1
-            print("No data from Powerparser.")
-            output, error_flag = ({'unixtime':None, 
-                                   0:None, 1:None, 2:None, 3:None, 4:None, 5:None, 6:None, 7:None, 
-                                   8:None, 9:None, 10:None, 11:None, 12:None, 13:None, 14:None, 15:None},
-                                   None)
-        return output, error_flag
+        data, errors = rtdparser(file_raw=raw_data)
+        return data, errors
 
     def parsed_2_collection(self, parsed_data):
         """
@@ -108,9 +98,9 @@ class PowerReader(ReaderBase):
         """
         # take human readable and convert and set to 
         # CdTeCollection(), TimePixCollection(), CMOSCollection()
-        col = PowerCollection(parsed_data, self.old_data_time)
-        # if col.last_data_time>self.old_data_time:
-        #     self.old_data_time = col.last_data_time
-        # if not hasattr(self,"data_start_time"):
-        #     self.data_start_time = col.output['unixtime'][0]
+        col = RTDCollection(parsed_data, self.old_data_time)
+        if col.last_data_time>self.old_data_time:
+            self.old_data_time = col.last_data_time
+        if not hasattr(self,"data_start_time"):
+            self.data_start_time = col.event['ti'][0]
         return col
